chore: [#184814757] split shared onboarding and profile components
